+++ etl_framework/MySqlSchemaModifier.py
@@ -4,7 +4,7 @@
 #pylint: disable=too-many-arguments
 #pylint: disable=abstract-class-instantiated
 
-from etl_framework.datastores import MySqlDatabase
+from etl_framework.datastores.mysql_database import MySqlDatabase
 from loader_mixins.SetConfigMixin import SetConfigMixin
 
 class MySqlSchemaModifier(MySqlDatabase,
